\name{addVariable.ggobiDataset}
\alias{addVariable.ggobiDataset}
\alias{addVariable}
\title{Add variable}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Add variable to a ggobiDataset
}
\usage{addVariable.ggobiDataset(x, vals, name, ...)}
\arguments{
\item{x}{ggobiDataset}
\item{vals}{values to add}
\item{name}{name of column to add}
\item{...}{}
}

\details{}

\examples{}
\keyword{manip}
