\name{"as.data.frame.GGobiData"}
\alias{as.data.frame.GGobiData}
\alias{as.matrix.GGobiData}
\title{Conversion methods}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Convert a GGobiData to a regular R data.frame or matrix
}
\usage{"as.data.frame.GGobiData"(x, ...)}
\arguments{
\item{x}{GGobiData}
\item{...}{}
}

\details{}

\examples{}
\keyword{manip}
\keyword{internal}
