\name{"as.data.frame.ggobiDataset"}
\alias{as.data.frame.ggobiDataset}
\alias{as.matrix.ggobiDataset}
\title{Conversion methods}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Convert a ggobiDataset to a regular R data.frame or matrix
}
\usage{"as.data.frame.ggobiDataset"(x, ...)}
\arguments{
\item{x}{ggobiDataset}
\item{...}{}
}

\details{}

\examples{}
\keyword{manip}
\keyword{internal}
