\name{colorscheme}
\alias{colorscheme}
\title{Get active colour scheme}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Get name of the active colour scheme
}
\usage{colorscheme(x)}
\arguments{
\item{x}{GGobi object}
}



\examples{
if (interactive()) {

g <- ggobi(mtcars)
colorscheme(g)}}
\keyword{color}
