\name{connecting_edges}
\alias{connecting_edges}
\title{Get connecting edges}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Get actual edges from application of edges dataset to target dataset.
}
\usage{connecting_edges(x, y)}
\arguments{
\item{x}{target ggobi dataset}
\item{y}{ggobi dataset containing edges}
}




\keyword{manip}
