\name{create_time_edges_for_unit}
\alias{create_time_edges_for_unit}
\title{Create time edges for a unit}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Create time edges for a unit
}
\usage{create_time_edges_for_unit(d, time)}
\arguments{
\item{d}{data frame of observations from one unit}
\item{time}{time variable}
}

\details{}

\examples{}
\keyword{dynamic}
\keyword{internal}
