\name{dim.GGobiData}
\alias{dim.GGobiData}
\title{GGobiData dimensions}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Retrieve the dimension of a GGobiData
}
\usage{dim.GGobiData(x)}
\arguments{
\item{x}{dataset}
}

\details{}

\examples{g <- ggobi(mtcars)
dim(g[1])}
\keyword{attribute}
\keyword{internal}
