\name{dim.ggobiDataset}
\alias{dim.ggobiDataset}
\title{ggobiDataset dimensions}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Retrieve the dimension of an ggobi dataset
}
\usage{dim.ggobiDataset(x)}
\arguments{
\item{x}{dataset}
}

\details{}

\examples{}
\keyword{attribute}
\keyword{internal}
