\name{"dimnames<-.GGobiData"}
\alias{dimnames<-.GGobiData}
\title{Set dim names}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Set dim names for a GGobiData
}
\usage{\S3method{dimnames}{GGobiData}(x) <- value}
\arguments{
\item{x}{GGobiData}
\item{value}{new names}
}




\keyword{attribute}
\keyword{internal}
