\name{dimnames.GGobiData}
\alias{dimnames.GGobiData}
\title{Get dimension names}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Get row and column names for a GGobiData
}
\usage{dimnames.GGobiData(x)}
\arguments{
\item{x}{ggobiDataget}
}

\details{}

\examples{}
\keyword{attribute}
\keyword{internal}
