\name{dimnames.ggobiDataset}
\alias{dimnames.ggobiDataset}
\title{Get dimension names}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Get row and column names for a ggobiDataget
}
\usage{dimnames.ggobiDataset(x)}
\arguments{
\item{x}{ggobiDataget}
}

\details{}

\examples{}
\keyword{attribute}
\keyword{internal}
