\name{"edges<-.GGobi"}
\alias{edges<-.GGobi}
\title{Set edges}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Create a new edges dataset and add to GGobi
}
\usage{"edges<-.GGobi"(x, value)}
\arguments{
\item{x}{GGobi dataset}
\item{value}{new edges}
}

\details{@arguments GGobi dataset
@arguments new edges}


\keyword{dynamic}
