\name{"edges<-.GGobiDisplay"}
\alias{edges<-.GGobiDisplay}
\title{Set edges}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Set edges for a display
}
\usage{"edges<-.GGobiDisplay"(x, value)}
\arguments{
\item{x}{GGobiDisplay object}
\item{value}{GGobiData object that contains edges}
}

\details{This sets the dataset that a GGobiDisplay uses
to display edges.}

\examples{}
\keyword{dynamic}
