\name{"excluded<-.GGobiData"}
\alias{excluded<-.GGobiData}
\title{Set excluded status}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Set the exclusion status of points.
}
\usage{\S3method{excluded}{GGobiData}(x) <- value}
\arguments{
\item{x}{GGobiData}
\item{value}{logical vector}
}

\details{If a point is excluded it is not drawn.}
\seealso{\code{\link{excluded}}}

\keyword{dynamic}
