\name{"excluded<-.ggobiDataset"}
\alias{excluded<-.ggobiDataset}
\alias{excluded<-}
\title{Set excluded status}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Set the exclusion status of points.
}
\usage{"excluded<-.ggobiDataset"(x, value)}
\arguments{
\item{x}{ggobiDataset}
\item{value}{logical vector}
}

\details{If a point is excluded it is not drawn.}
\seealso{\code{\link{excluded}}}
\examples{}
\keyword{dynamic}
