\name{excluded.GGobiData}
\alias{excluded.GGobiData}
\title{Get excluded status}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Get the exclusion status of points.
}
\usage{excluded.GGobiData(x)}
\arguments{
\item{x}{ggobiDataget}
}

\details{If a point is excluded it is not drawn.}
\seealso{\code{\link{excluded<-}}}
\examples{}
\keyword{dynamic}
