\name{ggobi_data_add_variable}
\alias{ggobi_data_add_variable}
\alias{addVariable}
\title{Add variable}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Add variable to a GGobiData
}
\usage{ggobi_data_add_variable(x, vals, name, ...)}
\arguments{
\item{x}{GGobiData}
\item{vals}{values to add}
\item{name}{name of column to add}
\item{...}{}
}

\details{}

\examples{}
\keyword{internal}
