\name{ggobi_data_remove_variable}
\alias{ggobi_data_remove_variable}
\title{}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
@keyword internal
}
\usage{ggobi_data_remove_variable(x, var)}
\arguments{
\item{x}{}
\item{var}{}
}

\details{@keyword internal}

\examples{}
\keyword{internal}
