\name{ggobi_data_set_variable}
\alias{ggobi_data_set_variable}
\title{Set variable values}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Set the variable values for a column in a GGobiData
}
\usage{ggobi_data_set_variable(x, vals, var, update = TRUE)}
\arguments{
\item{x}{GGobiData}
\item{vals}{values of new variable}
\item{var}{variable name}
\item{update}{update?}
}




\keyword{internal}
