\name{ggobi_data_set_variables}
\alias{ggobi_data_set_variables}
\title{Set variable values}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Set the variable values for a column in a ggobiDataset
}
\usage{ggobi_data_set_variables(x, values, var, update = TRUE)}
\arguments{
\item{x}{ggobiDataset}
\item{values}{values of new variable}
\item{var}{variable name}
\item{update}{update?}
}

\details{}

\examples{}
\keyword{manip}
\keyword{internal}
