\name{ggobi_data_write_xml}
\alias{ggobi_data_write_xml}
\title{Write xml}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Write GGobi xml for specific dataset to filename
}
\usage{ggobi_data_write_xml(gd, filename)}
\arguments{
\item{gd}{GGobiData object}
\item{filename}{path to write file to}
}

\details{}

\examples{}
\keyword{manip}
