\name{ggobi_display_get_plot_range}
\alias{ggobi_display_get_plot_range}
\title{Get plot range}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Get plot range as list of X and Y ranges
}
\usage{ggobi_display_get_plot_range(gd, plot=1)}
\arguments{
\item{gd}{GGobiDisplay object}
\item{plot}{plot number}
}

\details{}

\examples{}
\keyword{internal}
