\name{ggobi_display_make_type}
\alias{ggobi_display_make_type}
\title{Convert plot name to GGobi plot type}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Used to convert between friendly plot name and internal GGobi name.
}
\usage{ggobi_display_make_type(type)}
\arguments{
\item{type}{}
}

\details{}

\examples{}
\keyword{internal}
