\name{ggobi_display_set_plot_range}
\alias{ggobi_display_set_plot_range}
\title{Set plot range}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Set the range of a plot.
}
\usage{ggobi_display_set_plot_range(gd, plot=1, x, y)}
\arguments{
\item{gd}{GGobiDisplay object}
\item{plot}{plot number}
\item{x}{x range}
\item{y}{y range}
}

\details{Only applies to scatterplots.}

\examples{}
\keyword{internal}
