\name{ggobi_display_types}
\alias{ggobi_display_types}
\title{GGobiDisplay types}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Get list of GGobiDisplay types.  An instance of GGobi must be open
}
\usage{ggobi_display_types()}
\arguments{
\item{}{GGobi reference}
}


\seealso{\code{\link{ggobi_display_make_type}}}

\keyword{internal}
