\name{ggobi_find_file}
\alias{ggobi_find_file}
\title{Find GGobi file.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Finds a file stored somewhere in the ggobi installation.
}
\usage{ggobi_find_file(..., check = F)}
\arguments{
\item{...}{bits of the path to join together}
\item{check}{}
}

\details{}

\examples{ggobi_find_file("data","tips.xml")}
\keyword{dynamic}
\keyword{internal}
