\name{ggobi.GGobiDisplay}
\alias{ggobi.GGobiDisplay}
\title{Get display GGobi}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Returns a link to the GGobi object associated with this display
}
\usage{\S3method{ggobi}{GGobiDisplay}(data, \dots)}
\arguments{
\item{data}{GGobiDisplay object}
\item{\dots}{}
}




\keyword{internal}
