\name{ggobi_gtk_main_window}
\alias{ggobi_gtk_main_window}
\title{Gtk main window}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Retrieve RGtk object for main window
}
\usage{ggobi_gtk_main_window(.gobi = ggobi_get())}
\arguments{
\item{.gobi}{}
}

\details{Useful for embedding in other applications or for listening
to their signals via RGtk2.}

\examples{}
\keyword{internal}
