\name{ggobi_gtk_menu_bar}
\alias{ggobi_gtk_menu_bar}
\title{Gtk menu bar}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Retrieve RGtk object for menu bar
}
\usage{ggobi_gtk_menu_bar(.gobi = ggobi_get())}
\arguments{
\item{.gobi}{}
}

\details{Useful for embedding in other applications or for listening
to their signals via RGtk2.}


\keyword{internal}
