\name{ggobi_pcp}
\alias{ggobi_pcp}
\title{Create parallel coordinates plot.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Mock up a pcp plot using points and edges.
}
\usage{ggobi_pcp(data, type="range")}
\arguments{
\item{data}{}
\item{type}{}
}

\details{Experimental and may suggest ways to reduce PCP code in GGobi}

\examples{}
\keyword{internal}
