\name{ggobi_set_data_file}
\alias{ggobi_set_data_file}
\title{Set data file.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Open data file and add to ggobi datasets.
}
\usage{ggobi_set_data_file(file, mode = "unknown", add = TRUE, .gobi = ggobi_get())}
\arguments{
\item{file}{path to file}
\item{mode}{mode of file}
\item{add}{add?}
\item{.gobi}{ggobi instance}
}

\details{}

\examples{}
\keyword{manip}
\keyword{internal}
