\name{"glyph_colour<-.GGobiData"}
\alias{glyph_colour<-.GGobiData}
\alias{glyph_colour<-}
\alias{glyph_color<-}
\title{Set glyph colour}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Set glyph colour
}
\usage{\S3method{glyph_colour}{GGobiData}(x) <- value}
\arguments{
\item{x}{GGobiData}
\item{value}{vector of new colours}
}


\seealso{\code{\link{glyph_colour}}}

\keyword{dynamic}
