\name{glyph_colour.GGobiData}
\alias{glyph_colour.GGobiData}
\alias{glyph_colour}
\alias{glyph_color}
\title{Get glyph colour}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Get glyph colour
}
\usage{\S3method{glyph_colour}{GGobiData}(x)}
\arguments{
\item{x}{GGobiData}
}


\seealso{\code{\link{glyph_colour<-}}}

\keyword{dynamic}
