\name{glyph_colour.ggobiDataset}
\alias{glyph_colour.ggobiDataset}
\alias{glyph_colour}
\alias{glyph_color}
\title{Get glyph colour}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Get glyph colour
}
\usage{glyph_colour.ggobiDataset(x)}
\arguments{
\item{x}{ggobiDataset}
}

\details{}
\seealso{\code{\link{glyph_colour<-}}}
\examples{}
\keyword{dynamic}
