\name{"glyph_size<-.GGobiData"}
\alias{glyph_size<-.GGobiData}
\alias{glyph_size<-}
\title{Set glyph size}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Set glyph size
}
\usage{"glyph_size<-.GGobiData"(x, value)}
\arguments{
\item{x}{GGobiData}
\item{value}{vector of new sizes}
}

\details{Glyph size is an integer between 1 and 6.}
\seealso{\code{\link{glyph_size}}}
\examples{}
\keyword{dynamic}
