\name{"glyph_size<-.ggobiDataset"}
\alias{glyph_size<-.ggobiDataset}
\alias{glyph_size<-}
\title{Set glyph size}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Set glyph size
}
\usage{"glyph_size<-.ggobiDataset"(x, value)}
\arguments{
\item{x}{ggobiDataset}
\item{value}{vector of new sizes}
}

\details{}
\seealso{\code{\link{glyph_size}}}
\examples{}
\keyword{dynamic}
