\name{glyph_size.GGobiData}
\alias{glyph_size.GGobiData}
\alias{glyph_size}
\title{Get glyph size}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Get glyph size
}
\usage{glyph_size.GGobiData(x, value)}
\arguments{
\item{x}{GGobiData}
\item{value}{}
}

\details{Glyph size is an integer between 1 and 6.}
\seealso{\code{\link{glyph_size<-}}}
\examples{}
\keyword{dynamic}
