\name{glyph_size.ggobiDataset}
\alias{glyph_size.ggobiDataset}
\alias{glyph_size}
\title{Get glyph size}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Get glyph size
}
\usage{glyph_size.ggobiDataset(x, value)}
\arguments{
\item{x}{ggobiDataset}
\item{value}{}
}

\details{}
\seealso{\code{\link{glyph_size<-}}}
\examples{}
\keyword{dynamic}
