\name{"glyph_type<-.ggobiDataset"}
\alias{glyph_type<-.ggobiDataset}
\alias{glyph_type<-}
\title{Set glyph type}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Set glyph type
}
\usage{"glyph_type<-.ggobiDataset"(x, value)}
\arguments{
\item{x}{ggobiDataset}
\item{value}{vector of new types}
}

\details{}
\seealso{\code{\link{glyph_type}}}
\examples{}
\keyword{dynamic}
