\name{glyph_type.GGobiData}
\alias{glyph_type.GGobiData}
\alias{glyph_type}
\title{Get glyph type.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Get glyph type.
}
\usage{glyph_type.GGobiData(x)}
\arguments{
\item{x}{GGobiData}
}

\details{Glyph type refers to the shape of the glyph, one of:

\itemize{
\item a filled circle
\item an empty circle
\item a filled square
\item an empty square
\item a single pixel
\item a plus sign
\item a cross
}}
\seealso{\code{\link{glyph_type<-}}}
\examples{}
\keyword{dynamic}
