\name{glyph_type.ggobiDataset}
\alias{glyph_type.ggobiDataset}
\alias{glyph_type}
\title{Get glyph type.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Get glyph type.
}
\usage{glyph_type.ggobiDataset(x)}
\arguments{
\item{x}{ggobiDataset}
}

\details{}
\seealso{\code{\link{glyph_type<-}}}
\examples{}
\keyword{dynamic}
