\name{"ids<-.ggobiDataset"}
\alias{ids<-.ggobiDataset}
\alias{ids<-}
\title{Set row ids}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Set row ids from a ggobiDataset
}
\usage{"ids<-.ggobiDataset"(x, value)}
\arguments{
\item{x}{ggobiDataset}
\item{value}{new values}
}

\details{}
\seealso{\code{\link{ids}}}
\examples{}
\keyword{manip}
