\name{ids.ggobiDataset}
\alias{ids.ggobiDataset}
\alias{ids}
\title{Row ids}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Retrive row ids from a ggobiDataset
}
\usage{ids.ggobiDataset(x)}
\arguments{
\item{x}{ggobiDataset}
}

\details{}
\seealso{\code{\link{ids<-}}}
\examples{}
\keyword{manip}
