\name{length.GGobiDisplay}
\alias{length.GGobiDisplay}
\title{Length method for GGobiDisplay}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Returns the number of plots within a given display
}
\usage{length.GGobiDisplay(gd)}
\arguments{
\item{gd}{GGobiDisplay object to close}
}

\details{}

\examples{}
\keyword{internal}
