\name{mapGlyphType}
\alias{mapGlyphType}
\title{Map glyph type.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Map glyph character code to number
}
\usage{mapGlyphType(types)}
\arguments{
\item{types}{vector of glyph character codes}
}

\details{}

\examples{}
\keyword{dynamic}
\keyword{internal}
