\name{"names<-.GGobiData"}
\alias{names<-.GGobiData}
\title{Set column names}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Set column names for a GGobiData
}
\usage{"names<-.GGobiData"(x, value)}
\arguments{
\item{x}{GGobiData}
\item{value}{new names}
}

\details{}

\examples{}
\keyword{attribute}
\keyword{internal}
