\name{"names<-.ggobiDataset"}
\alias{names<-.ggobiDataset}
\title{Set column names}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Set column names for a ggobiDataset
}
\usage{"names<-.ggobiDataset"(x, value)}
\arguments{
\item{x}{ggobiDataset}
\item{value}{new names}
}

\details{}

\examples{}
\keyword{attribute}
\keyword{internal}
