\name{names.GGobiData}
\alias{names.GGobiData}
\title{GGobiData column names}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Get column names for a GGobiData
}
\usage{\S3method{names}{GGobiData}(x, \dots)}
\arguments{
\item{x}{dataset}
\item{\dots}{}
}




\keyword{attribute}
\keyword{internal}
