\name{names.ggobiDataset}
\alias{names.ggobiDataset}
\title{ggobiDataset column names}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Get column names for a ggobiDataset
}
\usage{names.ggobiDataset(x, ...)}
\arguments{
\item{x}{dataset}
\item{...}{}
}

\details{}

\examples{}
\keyword{attribute}
\keyword{internal}
