\name{ncol.GGobiData}
\alias{ncol.GGobiData}
\title{GGobiData columns}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Retrieve the number of columns in a GGobiData
}
\usage{ncol.GGobiData(d)}
\arguments{
\item{d}{dataset}
}

\details{}

\examples{}
\keyword{attribute}
\keyword{internal}
