\name{ncol.ggobiDataset}
\alias{ncol.ggobiDataset}
\title{ggobiDataset columns}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Retrieve the number of columns in an ggobi dataset
}
\usage{ncol.ggobiDataset(d)}
\arguments{
\item{d}{dataset}
}

\details{}

\examples{}
\keyword{attribute}
\keyword{internal}
