\name{nrow.ggobiDataset}
\alias{nrow.ggobiDataset}
\title{ggobiDataset rows}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Retrieve the number of row in an ggobi dataset
}
\usage{nrow.ggobiDataset(d)}
\arguments{
\item{d}{dataset}
}

\details{}

\examples{}
\keyword{attribute}
\keyword{internal}
