\name{pmodes}
\alias{pmodes}
\alias{imodes}
\alias{pmodes.default}
\alias{pmodes.GGobiDisplay}
\alias{imodes.GGobiDisplay}
\alias{pmodes.character}
\alias{imodes.character}
\title{List available projection/interaction modes}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
These generic methods retrieves possible projection/interaction modes.
}
\usage{pmodes(x=NULL)}
\arguments{
\item{x}{}
}

\details{The default methods retrieve a list modes for all display types,
the character methods retrieve modes for a given display type, and
the GGobiDisplay methods return possible modes for a given display}

\examples{}
\keyword{internal}
