\name{print.GGobi}
\alias{print.GGobi}
\title{Print ggobi}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Prints summary of ggobi object by instance
}
\usage{print.GGobi(x, ...)}
\arguments{
\item{x}{ggobi object}
\item{...}{}
}


\seealso{\code{\link{summary.GGobi}}}

\keyword{dynamic}
\keyword{internal}
