\name{print.GGobiData}
\alias{print.GGobiData}
\title{Print GGobiData}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Print GGobiData
}
\usage{\S3method{print}{GGobiData}(x, \dots)}
\arguments{
\item{x}{GGobi dataset to retrieve}
\item{\dots}{}
}

\details{By default printing a GGobiData acts like
printing an R data.frame - ie. show all the data}


\keyword{attribute}
\keyword{internal}
