\name{print.ggobiDataset}
\alias{print.ggobiDataset}
\title{Print ggobiDataset}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Print ggobiDataset
}
\usage{print.ggobiDataset(x, ...)}
\arguments{
\item{x}{GGobi dataset to retrieve}
\item{...}{}
}

\details{By default printing a ggobiDataset acts like
printing an R data.frame - ie. show all the data}

\examples{}
\keyword{attribute}
\keyword{internal}
