\name{printGGobiGGobi}
\alias{printGGobiGGobi}
\title{Print ggobi}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Prints summary of ggobi object by instance
}
\usage{printGGobiGGobi(x, ...)}
\arguments{
\item{x}{ggobi object}
\item{...}{}
}

\details{}
\seealso{\code{\link{summary.ggobi}}}
\examples{}
\keyword{dynamic}
\keyword{internal}
