\name{"rownames<-.ggobiDataset"}
\alias{rownames<-.ggobiDataset}
\title{Set row names}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Set row names for a ggobiDataset
}
\usage{"rownames<-.ggobiDataset"(x, value)}
\arguments{
\item{x}{ggobiDataset}
\item{value}{new names}
}

\details{}

\examples{}
\keyword{attribute}
\keyword{internal}
