\name{rownames.GGobiData}
\alias{rownames.GGobiData}
\title{Get row names}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Get row names for a GGobiData
}
\usage{\S3method{rownames}{GGobiData}(x)}
\arguments{
\item{x}{ggobiDataget}
\item{}{new names}
}




\keyword{attribute}
\keyword{internal}
