\name{rownames.ggobiDataset}
\alias{rownames.ggobiDataset}
\title{Get row names}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Get row names for a ggobiDataget
}
\usage{rownames.ggobiDataset(x)}
\arguments{
\item{x}{ggobiDataget}
\item{}{new names}
}

\details{}

\examples{}
\keyword{attribute}
\keyword{internal}
