\name{selected.ggobiDataset}
\alias{selected.ggobiDataset}
\alias{selected}
\title{Get selection status}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Returns logical vector indicating if each point is under the brush
}
\usage{selected.ggobiDataset(x)}
\arguments{
\item{x}{ggobiDataset}
\item{}{logical vector}
}

\details{}

\examples{}
\keyword{dynamic}
