\name{"shadowed<-.ggobiDataset"}
\alias{shadowed<-.ggobiDataset}
\alias{shadowed<-}
\title{Set shadowed status}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Set the exclusion status of points.
}
\usage{"shadowed<-.ggobiDataset"(x, value)}
\arguments{
\item{x}{ggobiDataset}
\item{value}{logical vector}
}

\details{If a point is shadowed it is drawn in a dark gray colour.}
\seealso{\code{\link{shadowed}}}
\examples{}
\keyword{dynamic}
