\name{shadowed.GGobiData}
\alias{shadowed.GGobiData}
\alias{shadowed}
\title{Get shadowed status}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Get the exclusion status of points.
}
\usage{shadowed.GGobiData(x)}
\arguments{
\item{x}{ggobiDataget}
}

\details{If a point is shadowed it is drawn in a dark gray colour,
behind all non-shadowed points.  It cannot be selected.}
\seealso{\code{\link{shadowed<-}}}

\keyword{dynamic}
