\name{shadowed.ggobiDataset}
\alias{shadowed.ggobiDataset}
\alias{shadowed}
\title{Get shadowed status}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Get the exclusion status of points.
}
\usage{shadowed.ggobiDataset(x)}
\arguments{
\item{x}{ggobiDataget}
}

\details{If a point is shadowed it is drawn in a dark gray colour.}
\seealso{\code{\link{shadowed<-}}}
\examples{}
\keyword{dynamic}
