\name{summary.GGobiData}
\alias{summary.GGobiData}
\title{Summarise GGobiData.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Summarise a GGobiData with dimensions, mode and variable names.
}
\usage{\S3method{summary}{GGobiData}(object, \dots)}
\arguments{
\item{object}{GGobiData}
\item{\dots}{ignored}
}




\keyword{attribute}
