\name{summary.GGobiData}
\alias{summary.GGobiData}
\title{Summarise GGobiData.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Summarise a GGobiData with dimensions, mode and variable names.
}
\usage{summary.GGobiData(object, ...)}
\arguments{
\item{object}{GGobiData}
\item{...}{}
}

\details{}

\examples{}
\keyword{attribute}
