\name{summary.ggobiDataset}
\alias{summary.ggobiDataset}
\title{Summarise ggobiDataset.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Summarise a ggobiDataset with dimensions, mode and variable names.
}
\usage{summary.ggobiDataset(object, ...)}
\arguments{
\item{object}{ggobiDataset}
\item{...}{}
}

\details{}

\examples{}
\keyword{attribute}
