\name{summaryGGobiGGobi}
\alias{summaryGGobiGGobi}
\title{GGobi summary}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Get a description of the global state of the GGobi session.
}
\usage{summaryGGobiGGobi(object, ...)}
\arguments{
\item{object}{ggobi object}
\item{...}{}
}

\details{}

\examples{g <- ggobi(mtcars)
summary(g)}
\keyword{dynamic}
