\name{variable_index}
\alias{variable_index}
\title{Variable index}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Return indices corresponding to variable names
}
\usage{variable_index(x, names)}
\arguments{
\item{x}{GGobiData}
\item{names}{variable names}
}




\keyword{attribute}
\keyword{internal}
