\name{clean.ggobi}
\alias{clean.ggobi}
\title{Clean ggobi}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Clean arguments for ggobi
}
\usage{clean.ggobi(x)}
\arguments{
\item{x}{vector}
}

\details{Arguments for ggobi need to be in specific format.
This function helps ensure that.}

\examples{}
\keyword{dynamic}
\keyword{internal}
