\name{ggobi_data_remove_variable}
\alias{ggobi_data_remove_variable}
\title{Remove variable}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Remove variables from a GGobiData object
}
\usage{ggobi_data_remove_variable(x, var)}
\arguments{
\item{x}{}
\item{var}{}
}

\details{}

\examples{}
\keyword{internal}
