\name{ggobi.find.file}
\alias{ggobi.find.file}
\title{Find GGobi file.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Finds a file stored somewhere in the ggobi installation.
}
\usage{ggobi.find.file(..., check = F)}
\arguments{
\item{...}{bits of the path to join together}
\item{check}{}
}

\details{}

\examples{ggobi.find.file("data","tips.xml")}
\keyword{dynamic}
\keyword{internal}
