\name{glyph_size.GGobiData}
\alias{glyph_size.GGobiData}
\alias{glyph_size}
\title{Get glyph size}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Get glyph size
}
\usage{\S3method{glyph_size}{GGobiData}(x)}
\arguments{
\item{x}{GGobiData}
}

\details{Glyph size is an integer between 1 and 6.}
\seealso{\code{\link{glyph_size<-}}}

\keyword{dynamic}
