\name{"glyph_type<-.GGobiData"}
\alias{glyph_type<-.GGobiData}
\alias{glyph_type<-}
\title{Set glyph type}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Set glyph type
}
\usage{\S3method{glyph_type}{GGobiData}(x) <- value}
\arguments{
\item{x}{GGobiData}
\item{value}{vector of new types}
}


\seealso{\code{\link{glyph_type}}}

\keyword{dynamic}
