\name{"ids<-.GGobiData"}
\alias{ids<-.GGobiData}
\alias{ids<-}
\title{Set row ids}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Set row ids from a GGobiData
}
\usage{"ids<-.GGobiData"(x, value)}
\arguments{
\item{x}{GGobiData}
\item{value}{new values}
}

\details{}
\seealso{\code{\link{ids}}}
\examples{}
\keyword{manip}
