\name{length.GGobiDisplay}
\alias{length.GGobiDisplay}
\title{Length method for GGobiDisplay}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Returns the number of plots within a given display
}
\usage{length.GGobiDisplay(x)}
\arguments{
\item{x}{GGobiDisplay object from which to retrieve the number of plots}
}

\details{}

\examples{}
\keyword{internal}
