\name{print.GGobi}
\alias{print.GGobi}
\title{Print ggobi}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Prints summary of ggobi object by instance
}
\usage{\S3method{print}{GGobi}(x, \dots)}
\arguments{
\item{x}{ggobi object}
\item{\dots}{}
}


\seealso{\code{\link{summary.GGobi}}}

\keyword{dynamic}
\keyword{internal}
