\name{print.GGobiData}
\alias{print.GGobiData}
\title{Print GGobiData}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Print GGobiData
}
\usage{print.GGobiData(x, ...)}
\arguments{
\item{x}{GGobi dataset to retrieve}
\item{...}{}
}

\details{By default printing a GGobiData acts like
printing an R data.frame - ie. show all the data}


\keyword{attribute}
\keyword{internal}
