\name{print.GGobiDisplay}
\alias{print.GGobiDisplay}
\title{Print method for GGobiDisplay}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Shows display type (first element in class vector)
}
\usage{\S3method{print}{GGobiDisplay}(x, \dots)}
\arguments{
\item{x}{GGobiDisplay object}
\item{\dots}{}
}




\keyword{internal}
