\name{print.GGobiDisplay}
\alias{print.GGobiDisplay}
\title{Print method for GGobiDisplay}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Shows display type (first element in class vector)
}
\usage{print.GGobiDisplay(x, ...)}
\arguments{
\item{x}{GGobiDisplay object}
\item{...}{}
}

\details{}

\examples{}
\keyword{internal}
