\name{selected.GGobiData}
\alias{selected.GGobiData}
\alias{selected}
\title{Get selection status}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Returns logical vector indicating if each point is under the brush
}
\usage{selected.GGobiData(x)}
\arguments{
\item{x}{GGobiData}
\item{}{logical vector}
}




\keyword{dynamic}
