\name{summary.GGobi}
\alias{summary.GGobi}
\title{GGobi summary}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Get a description of the global state of the GGobi session.
}
\usage{\S3method{summary}{GGobi}(object, \dots)}
\arguments{
\item{object}{ggobi object}
\item{\dots}{ignored}
}



\examples{g <- ggobi(mtcars)
summary(g)}
\keyword{dynamic}
